+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Config option to allow repairs without repair kits (item.repairKits.efficiency.missing). Default value is zero, which requires repair kits. If enabled, repair kits may still be used. [#316]
 - Display a chat message on player login when part/material models fail to load properly, directing the player to the log file
 ### Fixed
 - Non-main parts not displaying stats in tooltip [#314]